+++ boilerplates/nuxt/docs/v0.0.1/changelog.md
@@ -6,6 +6,7 @@
 
 ## 📓 Docs:
 - changelog workflow created (on push to develop, master and on release), manual changelog specified version added
+- Accordion component created with test and stories
 
 ## 🧹 Chores: